Prerelease 5.1.0 > master

+++ geodezyx/operational/gins_runner/gins_bd_update.py
@@ -30,17 +30,85 @@ def read_credentials(file_path):
     return login, password
 
 
+# def download_rsync(
+#     file_list,
+#     source_path,
+#     local_destination,
+#     remote_user=None,
+#     remote_host=None,
+#     rsync_options=None,
+#     password=None,
+# ):
+#     """
+#     Downloads a list of files using rsync.
+#     """
+#     if not rsync_options:
+#         rsync_options = [
+#             "--archive",
+#             "--verbose",
+#             "--compress",
+#             "--progress",
+#             "--relative",
+#             "--copy-links",
+#             "--no-perms",
+#             "--no-owner",
+#             "--no-group",
+#             "--omit-dir-times",
+#             #"--keep-dirlinks",
+#         ]  # Default options: archive mode, verbose, and compression
+#
+#     # Construct the remote source path
+#     if remote_user and remote_host:
+#         remote_source = f"{remote_user}@{remote_host}:{source_path}"
+#     else:
+#         remote_source = f"{source_path}"
+#
+#     if password:
+#         rsync_base = ["sshpass", "-p", password, "rsync"] + rsync_options
+#     else:
+#         rsync_base = ["rsync"] + rsync_options
+#
+#     # if exclude_compress:
+#     #   '--ignore-existing' '--exclude-from=<(echo "$(find /destination -name '*.gz' | sed 's/\.gz$//')"')'
+#
+#     tmp_rsync_file_lis = f"/tmp/" + utils.get_timestamp() + "_tmp_rsync_file_list.lst"
+#     utils.write_in_file("\n".join(file_list), tmp_rsync_file_lis)
+#
+#     # Construct the rsync command
+#     # /./ : https://askubuntu.com/questions/552120/preserve-directory-tree-while-copying-with-rsync
+#     rsync_cmd = (
+#         rsync_base
+#         + ["--files-from", tmp_rsync_file_lis]
+#         + [f"{remote_source}/./", f"{local_destination}/"]
+#     )
+#     if False:
+#         log.info("Rsync command: %s", " ".join(rsync_cmd))
+#
+#     # Run the rsync command
+#     process = subprocess.run(rsync_cmd, stderr=sys.stderr, stdout=sys.stdout, text=True)
+#
+#     if process.returncode != 0:
+#         log.error("Rsync failed with return code :( %d", process.returncode)
+#     else:
+#         log.info("Rsync completed successfully :)")
+#
+#     os.remove(tmp_rsync_file_lis)
+#     return
+
+
 def download_rsync(
-    file_list,
-    remote_user,
-    remote_host,
-    remote_path,
-    local_destination,
+    file_list=None,
+    source_path="",
+    local_destination="",
+    remote_user=None,
+    remote_host=None,
     rsync_options=None,
     password=None,
 ):
     """
-    Downloads a list of files using rsync.
+    Downloads files using rsync.
+    If file_list is provided, downloads only those files.
+    If file_list is None, downloads all files and subfolders from source_path.
     """
     if not rsync_options:
         rsync_options = [
@@ -48,37 +116,43 @@ def download_rsync(
             "--verbose",
             "--compress",
             "--progress",
-            "--relative",
             "--copy-links",
             "--no-perms",
             "--no-owner",
             "--no-group",
             "--omit-dir-times",
-        ]  # Default options: archive mode, verbose, and compression
+        ]
 
     # Construct the remote source path
-    remote_source = f"{remote_user}@{remote_host}:{remote_path}"
+    if remote_user and remote_host:
+        remote_source = f"{remote_user}@{remote_host}:{source_path}"
+    else:
+        remote_source = f"{source_path}"
 
     if password:
         rsync_base = ["sshpass", "-p", password, "rsync"] + rsync_options
     else:
         rsync_base = ["rsync"] + rsync_options
 
-    # if exclude_compress:
-    #   '--ignore-existing' '--exclude-from=<(echo "$(find /destination -name '*.gz' | sed 's/\.gz$//')"')'
+    # Construct the rsync command based on whether file_list is provided
+    if file_list is not None:
+        rsync_options.append("--relative")
+        # Download specific files from the list
+        tmp_rsync_file_lis = f"/tmp/" + utils.get_timestamp() + "_tmp_rsync_file_list.lst"
+        utils.write_in_file("\n".join(file_list), tmp_rsync_file_lis)
 
-    tmp_rsync_file_lis = f"/tmp/" + utils.get_timestamp() + "_tmp_rsync_file_list.lst"
-    utils.write_in_file("\n".join(file_list), tmp_rsync_file_lis)
+        rsync_cmd = (
+            rsync_base
+            + ["--files-from", tmp_rsync_file_lis]
+            + [f"{remote_source}/./", f"{local_destination}/"]
+        )
+    else:
+        tmp_rsync_file_lis = None
+        # Download everything from source_path
+        rsync_cmd = rsync_base + [f"{remote_source}/*", f"{local_destination}/"]
 
-    # Construct the rsync command
-    # /./ : https://askubuntu.com/questions/552120/preserve-directory-tree-while-copying-with-rsync
-    rsync_cmd = (
-        rsync_base
-        + ["--files-from", tmp_rsync_file_lis]
-        + [f"{remote_source}/./", f"{local_destination}/"]
-    )
-
-    log.debug("Rsync command: %s", " ".join(rsync_cmd))
+    if True:
+        log.info("Rsync command: %s", " ".join(rsync_cmd))
 
     # Run the rsync command
     process = subprocess.run(rsync_cmd, stderr=sys.stderr, stdout=sys.stdout, text=True)
@@ -88,17 +162,21 @@ def download_rsync(
     else:
         log.info("Rsync completed successfully :)")
 
-    os.remove(tmp_rsync_file_lis)
+    # Clean up temp file if it was created
+    if file_list is not None:
+        os.remove(tmp_rsync_file_lis)
+
     return
 
-
 def bdgins_update(
-    date_srt=dt.datetime(2020, 5, 3),
+    date_srt=dt.datetime(2000, 5, 3),
     date_end=None,
     dir_bdgins="",
     login="",
     password="",
     compress=False,
+    rapid=False,
+    ultra=False,
 ):
     """
     Update the BDGINS repository with the necessary files
@@ -122,6 +200,10 @@ def bdgins_update(
         Whether to gzip-compress the clock files.
         experimental, and not recommended
         Defaults to False.
+    rapid : bool, optional
+        download RAPID products. Defaults to False.
+    ultra : bool, optional
+        download ULTRA products. Defaults to False.
     
     Returns
     -------
@@ -162,33 +244,54 @@ def bdgins_update(
     ### time dependant files
     list_tropo_vmf1 = []
     list_iono_igs = []
-    list_orbite_mg3 = []
 
-    list_orbite_g20 = []
+    # list_orbite_sp3_mg3 = [] # not needed after v25_1
+    list_orbite_sp3_grg = []
+    list_orbite_sp3_g20 = []
+
+    list_orbite_gin_g20 = []
     list_orbex_g20 = []
     list_horl_g20 = []
+
+    list_all_grr = []
+    list_orbite_gin_grr = []
+    list_orbex_grr = []
+    list_horl_grr = []
+    list_clk_grr = []
+    list_bias_grr = []
+    list_erp_grr = []
+    list_prob_grr = []
+    list_nar_grr = []
+    list_sum_grr = []
+    list_gin_grr = []
+
+    list_all_gru = []
+    list_orbite_gin_gru = []
+    list_orbex_gru = []
+    list_horl_gru = []
+    list_clk_gru = []
+    list_bias_gru = []
+    list_erp_gru = []
+    list_prob_gru = []
+    list_nar_gru = []
+    list_sum_gru = []
+    list_gin_gru = []
 
     ###### LIST FILL
     ### full folders
     # (folder's path is added in the rsync command, with subdir destination variable
-    ## list_antex.extend(["/"]) ## temporary off because updated continuously
-    ## list_constell.extend(["/"])
-    ## list_exe_ppp.extend(["/"])
-    ## list_lunisolaires.extend(["/"]) ## temporary off because updated continuously
-    ## list_macromod.extend(["/"])
-    ## list_maree_polaire.extend(["/"])
-    ## list_pole.extend(["/"])
-    list_prairie.extend(["/"])
+    # list_antex.extend(["igs20.atx",
+    #                    "igs05.atx",
+    #                    "igsR3_2077.atx",
+    #                    "igsR3_2194.atx",
+    #                    "igsR3.atx",
+    #                    "igs20.atx",
+    #                    "igsR3_2107.atx",
+    #                    "igsR3_2233.atx",
+    #                    ]) ## temporary off because updated continuously
 
-    ### misc files: the needed files are considered individually
-    # some of them are a redundancy since they must be downloaded in the full folders
-    list_misc.extend([f"prairie/igs_satellite_metadata.snx"])
-    list_misc.extend([f"pole/nominal_NRO"])
-    list_misc.extend([f"ANTEX/igs20.atx"])
-    list_misc.extend([f"EXE_PPP/valap_static"])
-    list_misc.extend([f"macromod/gnss.xml"])
-    list_misc.extend([f"lunisolaires/de440bdlf.ad"])
-    list_misc.extend([f"maree_polaire/loading/nominal"])
+    list_antex = None
+
     l_fil_cons = [
         "constellation_gps.infos",
         "histocom.infos",
@@ -196,13 +299,35 @@ def bdgins_update(
         "historik_glonass",
         "igs_satellite_metadata.snx",
     ]
-    list_misc.extend(["constell/" + f for f in l_fil_cons])
+    list_constell.extend(l_fil_cons)
+    list_exe_ppp.extend(["valap_static"])
+    list_lunisolaires.extend(["de440bdlf.ad"]) ## temporary off because updated continuously
+    list_macromod.extend(["gnss.xml"])
+    list_maree_polaire.extend(["nominal"])
+    list_pole.extend(["nominal_NRO", "liste_sauts_tucs_1s.dat"])
+    # list_prairie.extend(["igs_satellite_metadata.snx"])
+
+    ### misc files: the needed files are considered individually
+    # some of them are a redundancy since they must be downloaded in the full folders
+    # list_misc.extend([f"pole/nominal_NRO"])
+    # list_misc.extend([f"ANTEX/igs20.atx"])
+    # list_misc.extend([f"EXE_PPP/valap_static"])
+    # list_misc.extend([f"macromod/gnss.xml"])
+    # list_misc.extend([f"lunisolaires/de440bdlf.ad"])
+    # list_misc.extend([f"maree_polaire/loading/nominal"])
+    # list_misc.extend(["constell/" + f for f in l_fil_cons])
+    list_misc.extend([f"prairie/igs_satellite_metadata.snx"])
 
     # here is an hybrid tropo/misc file
     list_tropo_vmf1.append(f"orography_ell")
 
     ### time dependant files
-    while date <= date_end + dt.timedelta(days=2): # 2 days later is need for cat orb/clk
+    if rapid or ultra:
+        extra_days = dt.timedelta(days=0)
+    else:
+        extra_days = dt.timedelta(days=2)  # 2 days later is need for cat orb/clk
+
+    while date <= date_end + extra_days:
         day = str(date.day).zfill(2)
         month = str(date.month).zfill(2)
         year = date.year
@@ -220,32 +345,94 @@ def bdgins_update(
             ]
         )
         list_iono_igs.append(f"{year}/igsg{doy}0.{yy}i.Z")
-        list_orbite_mg3.append(f"mg3{wk}{wkday}.sp3.Ci9PAU")
-        list_orbite_g20.append(f"G20{wk}{wkday}.gin")
+        #### Needed for the PRAIRIE pre-processing
+        # list_orbite_sp3_mg3.append(f"mg3{wk}{wkday}.sp3.Ci9PAU") # not needed after v25_1
+        if wk < 2245:
+            list_orbite_sp3_g20.append(f"g20{wk}{wkday}.sp3.Ci3G20") # for the past, no more g20 after wk2245
+        else:
+            list_orbite_sp3_grg.append(f"grg{wk}{wkday}.sp3") # for the routine
+        #### Needed for the main GINS calculation
+        list_orbite_gin_g20.append(f"G20{wk}{wkday}.gin")
         list_orbex_g20.append(f"G20{wk}{wkday}.obx.gz")
         list_horl_g20.append(f"hogps_g20{wk}{wkday}")
+        #### Needed for the main GINS calculation - RAPID
+        if rapid:
+            list_orbite_gin_grr.append(f"grr{wk}{wkday}.gin.gz")
+            list_all_grr.extend(list_orbite_gin_grr)
+            list_orbex_grr.append(f"grr{wk}{wkday}.obx.gz")
+            list_all_grr.extend(list_orbex_grr)
+            list_horl_grr.append(f"hogps_grr{wk}{wkday}.gz")
+            list_all_grr.extend(list_horl_grr)
+            list_clk_grr.append(f"grr{wk}{wkday}.clk.gz")
+            list_all_grr.extend(list_clk_grr)
+            list_bias_grr.append(f"grr{wk}{wkday}.bia.gz")
+            list_all_grr.extend(list_bias_grr)
+            list_erp_grr.append(f"grr{wk}{wkday}.erp.gz")
+            list_all_grr.extend(list_erp_grr)
+            list_prob_grr.append(f"grr{wk}{wkday}.problemes.gz")
+            list_all_grr.extend(list_prob_grr)
+            list_nar_grr.append(f"grr{wk}{wkday}.nar.gz")
+            list_all_grr.extend(list_nar_grr)
+            list_sum_grr.append(f"grr{wk}{wkday}.sum.gz")
+            list_all_grr.extend(list_sum_grr)
+            list_gin_grr.append(f"grr{wk}{wkday}.gin.gz")
+            list_all_grr.extend(list_gin_grr)
+
+        if ultra:
+            for hh in ["00", "06", "12", "18"]:
+                list_orbite_gin_gru.append(f"gru{wk}{wkday}_{hh}.gin.gz")
+                list_orbex_gru.append(f"gru{wk}{wkday}_{hh}.obx.gz")
+                list_horl_gru.append(f"hogps_gru{wk}{wkday}_{hh}.gz")
+                list_clk_gru.append(f"gru{wk}{wkday}_{hh}.clk.gz")
+                list_bias_gru.append(f"gru{wk}{wkday}_{hh}.bia.gz")
+                list_erp_gru.append(f"gru{wk}{wkday}_{hh}.erp.gz")
+                list_prob_gru.append(f"gru{wk}{wkday}_{hh}.problemes.gz")
+                list_nar_gru.append(f"gru{wk}{wkday}_{hh}.nar.gz")
+                list_sum_gru.append(f"gru{wk}{wkday}_{hh}.sum.gz")
+                list_gin_gru.append(f"gru{wk}{wkday}_{hh}.gin.gz")
+            list_all_gru.extend(list_orbite_gin_gru)
+            list_all_gru.extend(list_orbex_gru)
+            list_all_gru.extend(list_horl_gru)
+            list_all_gru.extend(list_clk_gru)
+            list_all_gru.extend(list_bias_gru)
+            list_all_gru.extend(list_erp_gru)
+            list_all_gru.extend(list_prob_gru)
+            list_all_gru.extend(list_nar_gru)
+            list_all_gru.extend(list_sum_gru)
+            list_all_gru.extend(list_gin_gru)
+
         date += dt.timedelta(days=1)
+
+    list_orbite_g20 = list_orbite_gin_g20 + list_orbite_sp3_g20
 
     ###### DESTINATION FOLDERS
     dest_subdir_dic = {
         ### full folders
-        #"ANTEX": list_antex,
-        #"constell": list_constell,
-        #"EXE_PPP": list_exe_ppp,
-        #"lunisolaires": list_lunisolaires,
-        #"macromod": list_macromod,
-        #"maree_polaire": list_maree_polaire,
-        "prairie": list_prairie,
+        "ANTEX": list_antex,
+        "constell": list_constell,
+        "EXE_PPP": list_exe_ppp,
+        "lunisolaires": list_lunisolaires,
+        "macromod": list_macromod,
+        "maree_polaire/loading": list_maree_polaire,
+        "pole": list_pole,
+        #"prairie": list_prairie,
         ### misc files
         ".": list_misc,  ## for misc files, destination is in the input path (.)
         ### time dependant files
         "tropo_vmf1": list_tropo_vmf1,
         "ionosphere/igs": list_iono_igs,
-        "orbites/SP3/re3": list_orbite_mg3,
+        #"orbites/SP3/re3": list_orbite_sp3_mg3, not needed after v25_1
+        "orbites/SP3/igs": list_orbite_sp3_grg,
         "mesures/gps/orbites/G20": list_orbite_g20,
         "mesures/gps/orbex/G20": list_orbex_g20,
         "mesures/gps/horloges30/G20": list_horl_g20,
+        ### RAPID
     }
+
+    if rapid:
+        dest_subdir_dic["mesures/gps/orbites/GRR"] = list_all_grr
+    if ultra:
+        dest_subdir_dic["mesures/gps/orbites/GRU"] = list_all_gru
 
     create_dir(dir_bdgins, subdirs=dest_subdir_dic.keys())
     os.chdir(dir_bdgins)
@@ -259,12 +446,25 @@ def bdgins_update(
 
         download_rsync(
             files_list,
-            login,
-            "tite.get.obs-mip.fr",
             remote_subdir_fullpath,
             local_subdir_fullpath,
+            login,
+            "tite.get.obs-mip.fr",
             password,
         )
+
+    # quick and dirty move for the PRAIRIE SP3 (VALIDE_25_1)
+    dir_gin_data = os.path.join(gynscmn.get_gin_path(True), "data")
+    dir_gin_sp3_g20 = os.path.join(dir_bdgins, "mesures/gps/orbites/G20")
+    dir_gin_sp3_grg = os.path.join(dir_bdgins, "orbites/SP3/igs")
+    dir_prairie_sp3_final = os.path.join(dir_gin_data, "orbites/SP3/gin/batch/orbite/G20/")
+
+    utils.create_dir(dir_prairie_sp3_final)
+    for dir_gin_sp3 in [dir_gin_sp3_g20, dir_gin_sp3_grg]:
+        log.info("Moving PRAIRIE's SP3 to their final destination: %s > %s", dir_gin_sp3, dir_prairie_sp3_final)
+        files_list = utils.find_recursive(dir_gin_sp3, "*.sp3*")
+        files_list = [os.path.basename(e) for e in files_list]
+        download_rsync(files_list, dir_gin_sp3, dir_prairie_sp3_final)
 
     # compress the clock files (experimental, and not recommended)
     if compress:
@@ -295,7 +495,7 @@ def create_dir(parent_dir, subdirs):
 
     for subdir in subdirs:
         subdir_fullpath = os.path.join(parent_dir, subdir)
-        if not os.path.exists(subdir_fullpath):
+        if not os.path.exists(subdir_fullpath) and not os.path.islink(subdir_fullpath):
             os.makedirs(subdir_fullpath)
 
 
@@ -306,10 +506,10 @@ def main():
         "--date_srt",
         type=lambda s: conv.date_pattern_2_dt(s),
         required=False,
-        default=dt.datetime(2020, 5, 3),
+        default=dt.datetime(2000, 5, 3),
         help=(
             "Start date for the update in various formats. "
-            "Default is 2020-05-03 "
+            "Default is 2000-05-03 "
             "(origin of G20 products, end of GPS's Selective Availability)."
         ),
     )
@@ -356,12 +556,31 @@ def main():
         "-c",
         "--compress",
         help=(
-            "Flag to enable gzip compression of clock files. "
+            "Enable gzip compression of clock files. "
             "Defaults to `False` and is experimental."
         ),
         action="store_true",
         default=False,
     )
+
+    parser.add_argument(
+        "-r",
+        "--rapid",
+        help="Enable downloading RAPID products.",
+        action="store_true",
+        dest="rapid",
+        default=False,
+    )
+
+    parser.add_argument(
+        "-u",
+        "--ultra",
+        help="Enable downloading ULTRA products.",
+        action="store_true",
+        dest="ultra",
+        default=False,
+    )
+
     args = parser.parse_args()
 
     bdgins_update(
@@ -371,6 +590,8 @@ def main():
         password=args.password,
         dir_bdgins=args.dir_bdgins,
         compress=args.compress,
+        rapid=args.rapid,
+        ultra=args.ultra,
     )
